1. Add new object mapping, e.g. one-to-many 2. Add support for Java 8 Date time

+++ src/main/java/hk/edu/cuhk/itm/repository/AdministratorRepository.java
@@ -3,11 +3,8 @@ package hk.edu.cuhk.itm.repository;
 import hk.edu.cuhk.itm.model.Administrator;
 import org.springframework.data.repository.Repository;
 
-/**
- * Created by hzy on 22/2/2016.
- */
 @org.springframework.stereotype.Repository
-public interface AdministratorRepository extends Repository<Administrator, Long> {
+public interface AdministratorRepository extends Repository<Administrator, Integer> {
 
-    Administrator findOne(Long id);
+    Administrator findOne(Integer id);
 }
